files moved successfully

+++ moveFile.py
@@ -2,6 +2,7 @@ import os
 import shutil
 import csv
 from pathlib import Path
+from os.path import exists
 
 def readCSV(fileName):
   records = []
@@ -14,12 +15,6 @@ def readCSV(fileName):
 images = readCSV('ImageFamily.csv')
 print(images)
 
-image_folder = str(Path('C:', '/', 'NSCF', 'ImageMove'))
-print(image_folder)
-family_folder = "C:\NSCF\Families"
-#family_folder = str(Path('C:', '/', 'NSCF', 'Families'))
-print(family_folder)
-
 
 # iterate files
 for image in images:
@@ -27,33 +22,19 @@ for image in images:
     familyName =image['Family'] 
     print(imageName)
 #construct full file path
-    source = image_folder + '/' + imageName
-    #destination = os.path.join(family_folder, familyName)
-    destination = family_folder + '/' + familyName
-    print(destination)
+    image_folder = Path(r'C:\\', 'NSCF', 'ImageMove')
+    print(image_folder)
+    family_folder = Path(r"C:\\", 'NSCF', 'ImageMove', 'Families', familyName)
+    print(family_folder)
+    source = os.path.join(image_folder, imageName)
+    if exists(source):
+        print("source is:", source)
+        destination = os.path.join(family_folder, imageName)
+        print("destination is:", destination)
 
-#test if family folder exists - if not create and then move
-    # if not os.path.exists(destination):
-    #     os.mkdir(destination)
-#move file
-    shutil.move(source, destination) 
-    print('Moved:', image['Barcode'])
-
-#example code
-# path = os.path.join(parent_dir, new_dir)
-# if not os.path.exists(path):
-#    os.mkdir(path)
-# # check if file exist in destination
-# if os.path.exists(dst_folder + file_name):
-#     # Split name and extension
-#     data = os.path.splitext(file_name)
-#     only_name = data[0]
-#     extension = data[1]
-#     # Adding the new name
-#     new_base = only_name + '_new' + extension
-#     # construct full file path
-#     new_name = os.path.join(dst_folder, new_base)
-#     # move file
-#     shutil.move(src_folder + file_name, new_name)
-# else:
-#     shutil.move(src_folder + file_name, dst_folder + file_name)
+    #test if family folder exists - if not create and then move
+        if not os.path.exists(family_folder):
+            os.mkdir(family_folder)
+    #move file
+        shutil.move(source, destination) 
+        print('Moved:', image['Barcode'])
